+++ LetsOrganize/Controllers/MyListController.cs
@@ -1,17 +1,16 @@
-﻿
-using LetsOrganize.Entities;
-using LetsOrganize.Services;
+﻿using LetsOrganize.Interfaces;
+using LetsOrganize.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LetsOrganize.Controllers
 {
-    
     [ApiController]
     [Route("[controller]")]
+    [Authorize]
     public class MyListController : ControllerBase
     {
         private readonly IMyListService _myListService;
-
 
         public MyListController(IMyListService myListService)
         {
@@ -22,32 +21,62 @@ namespace LetsOrganize.Controllers
         public async Task<ActionResult> GetLists()
         {
             var lists = await _myListService.GetAllLists();
+
             return Ok(lists);
         }
-        [HttpGet("/list/{id}")]
+
+        [HttpGet("/lists/{id}")]
         public async Task<ActionResult> GetListById([FromRoute] int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("Id must be a positive number.");
+            }
+
             var list = await _myListService.GetListById(id);
+
             return Ok(list);
         }
 
-        [HttpPost]
-        public async Task<ActionResult> Create([FromBody]string name)
+        [HttpPost("/lists")]
+        public async Task<ActionResult> Create([FromBody] MyListDto newListDto)
         {
-            var newList = await _myListService.CreateList(name);
+            if (newListDto == null)
+            {
+                return BadRequest("Incorrect data entered");
+            }
 
-            return Ok(newList);
-            //return CreatedAtAction(nameof(GetListById), name, newList);
+            var newCreatedList = await _myListService.CreateList(newListDto);
+
+            return Created($"/lists/{newCreatedList}", null);
         }
-        [HttpPut("/list/{id}")]
-        public async Task<ActionResult> ChangeListName([FromRoute] int id, [FromBody] string name)
+
+        [HttpPut("/lists/{id}")]
+        public async Task<ActionResult> EditList([FromRoute] int id, [FromBody] MyListDto newListDto)
         {
-            var updatedList = await _myListService.ChangeListName(id, name);
+            if (id <= 0)
+            {
+                return BadRequest("Id must be a positive number.");
+            }
+
+            if (newListDto == null)
+            {
+                return BadRequest("Incorrect data entered");
+            }
+
+            var updatedList = await _myListService.EditList(id, newListDto);
+            
             return Ok(updatedList);
         }
-        [HttpDelete("/list/{id}")]
+
+        [HttpDelete("/lists/{id}")]
         public async Task<ActionResult> DeleteList([FromRoute] int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest("Id must be a positive number.");
+            }
+
             await _myListService.DeleteList(id);
 
             return Ok();
